🌟 [feat]: Add scripts to load company and companies
Added script to load a specific company and display its details. Also added a script to load multiple companies based on a search query and display their details. Made changes to the login script to handle unsuccessful login attempts and throw an error.

Changes Made:
- Added `300.LoadCompany.ps1` script to load a specific company and display its details
- Added `310.LoadCompanies.ps1` script to load multiple companies based on a search query and display their details
- Modified `ImportAndLogin.ps1` script to handle unsuccessful login attempts and throw an error

Squashed commit of the following:

- Added script to load a specific company and display its details
- Added script to load multiple companies based on a search query and display their details
- Modified login script to handle unsuccessful login attempts

🚀 Please review and merge these changes.